backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
boto3==1.40.13
pdfplumber==0.11.7
pymupdf==1.26.4
reportlab==4.4.3
httpx==0.28.1
openai==1.99.9
PyJWT==2.10.1
sentry-sdk==2.13.0
